// types.ts
export interface ComparisonData {
  name: string;
  reklama: number;
  pr: number;
}

export interface GrowthData {
  period: string;
  ishonch: number;
  loyaqlik: number;
}

export interface PerceptionData {
  name: string;
  value: number;
  color: string;
}

export interface SectionContent {
  id: string;
  title: string;
  content: string;
}